requirements.txt
fastapi
uvicorn
python-dotenv
loguru
openai
torch
transformers
pydantic
nest_asyncio
